Cleanup + test fixes

+++ programs/uxd/src/lib.rs
@@ -36,7 +36,7 @@ pub const MAX_MANGO_DEPOSITORIES_REDEEMABLE_SOFT_CAP: u64 = u64::MAX;
 pub const DEFAULT_MANGO_DEPOSITORIES_REDEEMABLE_SOFT_CAP: u64 = 10_000; // 10 Thousand redeemable UI units
 
 const BPS_POW: u8 = 4; // Raise a number to BPS_POW to get order of magnitude of
-pub const BPS_UNIT_CONVERSION: u64 = (10 as u64).pow(BPS_POW as u32);
+pub const BPS_UNIT_CONVERSION: u64 = (10u64).pow(BPS_POW as u32);
 
 const SOLANA_MAX_MINT_DECIMALS: u8 = 9;
 
@@ -180,7 +180,7 @@ pub mod uxd {
     ///  Beyond 90% the `MangoAccount` can be liquidated by other mango accounts.
     ///  (And borrows/withdraws are still disabled)
     ///
-    /// Note: 
+    /// Note:
     ///  As the funding rate care be either negative or positive, the insurance
     ///  is there as a buffer to ensure that redeemables can be swapped back
     ///  at all time (by unwinding the backing amount of delta neutral
@@ -370,7 +370,10 @@ pub mod uxd {
         ctx: Context<QuoteMintWithMangoDepository>,
         quote_amount: u64,
     ) -> Result<()> {
-        msg!("[quote_mint_with_mango_depository] quote_amount {}", quote_amount);
+        msg!(
+            "[quote_mint_with_mango_depository] quote_amount {}",
+            quote_amount
+        );
         instructions::quote_mint_with_mango_depository::handler(ctx, quote_amount)
     }
 
@@ -381,7 +384,10 @@ pub mod uxd {
         ctx: Context<QuoteRedeemFromMangoDepository>,
         redeemable_amount: u64,
     ) -> Result<()> {
-        msg!("[quote_redeem_from_mango_depository] redeemable_amount {}", redeemable_amount);
+        msg!(
+            "[quote_redeem_from_mango_depository] redeemable_amount {}",
+            redeemable_amount
+        );
         instructions::quote_redeem_from_mango_depository::handler(ctx, redeemable_amount)
     }
 
@@ -389,7 +395,10 @@ pub mod uxd {
         ctx: Context<SetMangoDepositoryQuoteMintAndRedeemFee>,
         quote_fee: u8,
     ) -> Result<()> {
-        msg!("[set_mango_depository_quote_mint_and_redeem_fee] quote_fee {}", quote_fee);
+        msg!(
+            "[set_mango_depository_quote_mint_and_redeem_fee] quote_fee {}",
+            quote_fee
+        );
         instructions::set_mango_depository_quote_mint_and_redeem_fee::handler(ctx, quote_fee)
     }
 }
